+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # DocuSign Node Client Changelog
 
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
+## [v4.10.1] - eSignature API v2-20.3.01
+### Fixed 
+- DCM-3866, Added support for updateBrandResourcesByContentType function to take in file to upload.
+- DCM-3369, Updated ApiClient to use an empty JSON object if the body is null.
+- DCM-4614, Fixed out of memory issue when deserializing large files.
+
 ## [v4.10.0] - eSignature API v2-20.3.00
 ### Changed
 - Added support for version v2-20.3.00 of the DocuSign eSignature API.